Race condition patch
There seems to be a race condition between `pyftpsync` writing the file and the `api` service picking up the file. At times it picks up the file when it is only partially written creating a corrupted file,
Added sleep(1) before the first `shutil.copy2` will reduce the chance the file is not completely written when the api picks it up.